feat(pr-baseline-2): fix behavior when multiple reporters have errors (#943)
#### Details

This PR sets up the `AllProgressReporter` to run callbacks from all sub-reporters even if the first in sequence fails. If multiple fail, it uses a new `AggregateError` to report that.

##### Motivation

Ensures that we fail builds correctly/with reasonable errors even if PR creator encounters a problem

##### Context

I considered using the `aggregate-error` package rather than rolling our own, but it was simple enough to implement one ourselves that it didn't seem worth the extra overhead of adding a new dependency. It also would have ended up hitting the same issue that we're currently seeing with `get-port`, since it's an ESM-only package.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [ ] (after PR created) The `Accessibility Checks (pull_request)` check should fail. All other checks should pass.