docs(ado-extension): add note to have yarn as a prerequisite (#1099)
Details
The ADO extension uses Yarn to install packages before running scans on websites. Some build agents come with Yarn pre-installed, as indicated in docs for Microsoft-hosted agents. This adds a note to indicate to users to add a step to install Yarn if it is not already included in their build agent.

Motivation
It was previously unclear to users if Yarn was needed for the extension.