chore(deps): replace verror with pony-cause to preempt future node compat issues (#1100)
#### Details

Today, our logger uses `node-verror` to represent an `Error` object that can track an underlying `cause`. It does this by exposing a `cause()` method. However, since verror was written, JavaScript's TC39 group has standardized a `cause` property (not method) on the base `Error` type, which is incompatible (`cause()` vs `cause`).

This doesn't cause any issues for us today, but could in the future if, for example, we ever tracked exceptions through a library (eg, app insights) which tried to read from `cause`. It also causes errors in TypeScript 4.6.2 (see build failures in #1080), which catches the incompatible types.

This PR addresses the issue by swapping out `node-verror` for a similar library called `pony-cause` which implements an `ErrorWithCause` class comparable to `VError` but with the standards-compliant `cause` property. It also provides a `stackWithCauses` helper that I've dropped into our logger which is cross-compatible with both `cause` and `cause()`, which is helpful because several underlying `accessibility-insights-scan` errors still use `VError` for now.

##### Motivation

* Resolves build failure in #1080
* Preempts potential future compat issues when we eventually update node versions

##### Context

We'll want to make a comparable change the service; I've let Maxim know about the service impact offline already

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [ ] (after PR created) The `Accessibility Checks (pull_request)` check should fail. All other checks should pass.