feat(ado-ext-telemetry): make output artifact naming/grouping more consistent with outputDir (#1140)
#### Details

Per discussion with @karanbirsingh, this PR adjusts the naming and grouping of the new inputs for automatic artifact uploading to be more consistent with the existing `outputDir` input. Previously, `outputDir` was hidden under "Advanced options" and the artifact upload inputs were named `artifactName` and `uploadResultsAsArtifact`. This PR:

* Groups the three related inputs together under a new `Output options` grouping, which appears expanded by default
* Renames `uploadResultsAsArtifact` to `uploadOutputArtifact`, to make it more apparent that it is working with the same output that `outputDir` is
* Renames `artifactName` to `outputArtifactName`, for consistency with `uploadOutputArtifact`

It adjusts all usage/tests/docs accordingly.

##### Motivation

Feature work

##### Context

Separated this out from a separate (upcoming) PR that will adjust the default `outputDir` behavior to send results to a temp directory by default

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1940706
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)