feat(ado-ext-telemetry): update outputDir to write to a temp dir by default (#1141)
#### Details

This PR updates the ADO `outputDir` input to be optional without a default, instead of defaulting to `_accessibility-reports`, and to create a uniquely generated temp directory to use for `outputDir` if none is specified.

This PR intentionally does not update the GitHub Action with the same behavior because we didn't implement the embedded artifact uploading there, and it doesn't make sense to use a temp dir without the automatic artifact uploading. I did put the interesting part of the temp file generation in the `shared` package - I think if we wanted to implement it in the action in the future, it might makes sense to move the current `AdoTaskConfig.getReportOutDir` implementation into the base `TaskConfig` class as a wrapper around private-protected `getRawOutputDir` methods overriden by the subclasses.

##### Motivation

Now that uploading the output as an artifact is default, embedded behavior, the `outputDir` is an implementation detail for most users. We noticed in #1135 that it felt awkward to have to specify an implementation detail just to run the task multiple times from one pipeline, so this PR updates the default behavior to avoid having to do that.

This is technically breaking, since a user could previously have had a publish step that relied on the default _accessibility-reports directory, so it's been documented in the v1 to v2 docs.

##### Context

I think it would be nice for us to detect and give a nice error in the case where `outputDir` is specified explicitly in a way which is not unique among multiple steps in the same pipeline (probably by detecting if it already exists and erroring if so). I've left that out of scope for this PR since I expect it might make sense to fold it into the other input validation stuff @ThanyaLeif is in progress on.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1940706
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)