feat(ado-ext-telemetry): update staging validation pipeline with v2 changes (#1135)
#### Details

Release 2.0.0 bumped the major version of the staging extension to 2.0.0, but our staging validation workflow is still running against version `@1`. This updates the validation template to match the updated number.

Several of the scenarios were impacted by breaking changes, and there were a few new scenarios that make sense for us to validate. This PR also updates the scenarios to be based on the new v2 functionality.

##### Motivation

Enable release validation for 2.0.0

##### Context

The syntax in the template was updated already in #1122

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: Story 1927731
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)