chore: run validation scenarios in canary and prod as well as staging (#1185)
#### Details

Today, we have a variety of staging validation scenarios, but we don't actually run them until we kick off a staging release. It would be nice to be able to validate particularly interesting/risky changes against the same scenarios in Canary without kicking off a staging release. This PR runs the same scenarios that we already run for staging releases on canary and prod release as well.

No GHA impact.

##### Motivation

See above

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [x] Described how this PR impacts both the ADO extension and the GitHub action