fix: browser resolution set by ai-scan is 800x600 (#1364)
#### Details

The browser resolution set by ai-scan is 800x600, this pull request
updates the set `scanResultData.browserResolution` to match. This value
is printed at the bottom of reports. This pull request will also add a
bullet point to each usage document to specify the viewport size.

![The screenshot is the footnote shown at the bottom of AI scans. It
reads "with a display resolution of 1920x1080" when it should say "with
a display resolution of
800x600"](https://user-images.githubusercontent.com/2180540/199054241-1b93501c-a418-44b3-9e58-08e50fadcfd3.png)



##### Motivation

<!-- This can be as simple as "addresses issue #123" -->

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

Is there a way to pull this from the ai-scan to prevent future
mismatches? It doesn't look likely. This is something we should consider
if we decide to create viewport inputs.


<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [x] Described how this PR impacts both the ADO extension and the
GitHub action