chore: Fix feature request issue template (#1373)
#### Details

I noticed the feature request issue template was not appearing as an
option to select. I think this is because there is a empty line at the
start of the markdown file.

##### Motivation

I would like to request a feature 😇

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: Fixes #0000
- [ ] Added relevant unit test for your changes. (`yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [ ] Ran precheckin (`yarn precheckin`)
- [ ] Described how this PR impacts both the ADO extension and the
GitHub action