chore: add release change log generation tool (#1381)
#### Details

This adds a script to generate a changelog during release. This copies
from the optional script we have for Accessibility Insights for Web
releases.

It auto-generates the tester, verified, dev, commit, pr, change
description, group, version, and date columns for all the changes
between the current release and the previous.

To run it, in the root of the action repo you run the following command,
for this example I'm using the previous two release tags:
`node ./tools/get-change-log-for-release.js --from v3.1.0 --to v3.1.1 `

This is the resulting file from that command:

[change-log.v3.1.0-v3.1.1.csv](https://github.com/microsoft/accessibility-insights-action/files/9933118/change-log.v3.1.0-v3.1.1.csv)

Once the table is generated and filtered, it looks like this (an example
of what the previous release's table would have looked like):

![image](https://user-images.githubusercontent.com/13286385/199837391-68264996-e358-4d8b-97a1-5485fef626b8.png)

##### Motivation

Additional tooling to simplify the release process

##### Context

This will work the same for both the GH action and the ADO Extension.

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [n/a] Ran precheckin (`yarn precheckin`)
- [x] Described how this PR impacts both the ADO extension and the
GitHub action