fix: prevent e2e stdout from triggering build failure on Azure pipeline (#1417)
#### Details

I noticed that the e2e tests on the ADO pipeline are failing. I believe
this is because when we log the stdout from the e2e tests, the errors
collected from the tests are triggering failures in the pipeline. This
pull request will add a small function that will replace the ADO logging
commands with `##[error]` which will format as an error in the pipeline,
but not fail the pipeline.

Example failing build:
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=39572&view=logs&j=81b7b1a0-9e2e-58a5-1601-69aaad9b82d6&t=5cf72e78-07c0-5125-1262-b1e05b2475fe

##### Motivation

Unblock pipeline.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

For visibility, it would be ideal to keep printing the stdout with each
test, but I'm not sure how to prevent the collected errors from pulling
through. I tried:

- [x] Wrapping the stdout in a LogIssue warning logging command:
https://learn.microsoft.com/en-us/azure/devops/pipelines/scripts/logging-commands?view=azure-devops&tabs=bash#logissue-log-an-error-or-warning
-- didn't work
- [x] LogDetail logging command:
https://learn.microsoft.com/en-us/azure/devops/pipelines/scripts/logging-commands?view=azure-devops&tabs=bash#logdetail-create-or-update-a-timeline-record-for-a-task
-- didn't work
- [x] Use a string replacement to change errors to warnings -- worked!

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [x] Described how this PR impacts both the ADO extension and the
GitHub action