feat(decommission-gh-action): Remove action specific code from shared package (#1444)
#### Details

This pull request removes action specific code from the shared package:

- Remove execution environment from markdown reporters

##### Motivation

Feature work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

We considered keeping the execution environment, but it's only used for
the logic of switching the emoji. Since it would be fairly quick to add
back, we decided to remove the execution environment for now.


<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [n/a] Described how this PR impacts both the ADO extension and the
GitHub action