feat(decommission-gh-action): update documentation removing references to the GitHub Action (#1447)
#### Details

This updates all documentation within the repo to remove references to
the GitHub Action and adjust language to apply to the new situation of
the ADO Extension being the only product in the repo.

##### Motivation

Feature work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran prechecking (`yarn prechecking`)
- [x] Described how this PR impacts both the ADO extension and the
GitHub action