feat(decommission-gh-action): Remove gh-action from build pipeline (#1448)
#### Details

This pull request will remove the GitHub action from the build pipeline.

##### Motivation

<!-- This can be as simple as "addresses issue #123" -->

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

From what I understand, the gh-action was placed in an artifact at
`drop/packages/gh-action/`, while the extension lives in an artifact at
`ado-extension-drop/pkg/`. This PR will remove the `drop/` directory at
the root since that appears to be for the action only. The
REALEASE_COMMIT.md was saved in `ado-extension-drop/drop`, but I moved
that up a directory into `ado-extension-drop` to simplify the directory
structure.

- Example of artifact before this PR:
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=40217&view=artifacts&pathAsName=false&type=publishedArtifacts
- Example of artifact with this PR:
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=40411&view=artifacts&pathAsName=false&type=publishedArtifacts


<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [x] Described how this PR impacts both the ADO extension and the
GitHub action