chore: Update QueryAzureDevOpsExtensionVersion@4 (#1455)
#### Details

This PR will use the latest version of QueryAzureDevOpsExtensionVersion.
We are noticing the canary pipeline is not querying the correct version
number, so we will see if the latest version of
QueryAzureDevOpsExtensionVersion will help here.

##### Motivation

Unblock pipeline.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [na/] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)