feat(telemetry-and-inputs): add failOnAccessibilityError telemetry (#1509)
#### Details

This adds telemetry tracking of usage of the `failOnAccessibilityError`
input.

##### Motivation

To allow for visibility into compliance of failing pull requests if
accessibility errors are found and no baseline is in place. If
`failOnAccessibilityError` is set to `true`, the extension will always
return as successful even if there are errors, rendering the extension
mostly useless.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran prechecking (`yarn precheckin`)