feat(telemetry-and-inputs): add staticSiteDir telemetry (#1538)
#### Details

This adds staticSiteDir to the list of inputs telemetry that we're
collecting.

Rather than collect the entire value for `staticSiteDir`, this records a
`staticSiteDirSet` boolean value, as we don't need to know the path,
only if they're using the input or not.

##### Motivation

We want to track how often the static site option is being used with the
end goal of removing it in the future if it is not being used much.
Removing this input in the future would allow us to also remove the
`dynamicSite` and `staticSite` options that tend to cause confusion for
many teams onboarding. It also wouldn't prevent people from scanning
static sites, they would just need to serve those sites in the pipeline
on their own.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran prechecking (`yarn prechecking`)