feat(faq-update): point helpMarkdown to new FAQ link (#1568)
#### Details

This points the top-level `helpMarkdown` in the `task.json` file to the
new FAQ link. This snippet is displayed at the top of the task whenever
it is run.

Here is a run with my test extension displaying the new behavior:
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=42652&view=logs&j=12f1170f-54f2-53f3-20dd-22fc7dff55f9&t=7384d774-f7ca-599c-ee57-ab2c05be9247

##### Motivation

This allows us to point users to a single place for both configuration
and troubleshooting questions.

##### Context

We'll also be updating the eng.ms documentation to link back to setup
documentation.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran prechecking (`yarn precheckin`)