feat(onboard-1es-pt): Onboard to 1ES Pipeline Templates (#1610)
#### Details

Onboards the CI/build pipeline to 1ES Pipeline Templates.

- I was unable to onboard the pipeline to use the SDL TSA option. TSA v2
requires that a configuration file. In v1, the pipeline passes inputs
and stored the values as pipeline variables. While the pipeline can
generate the TSA configuration file on the pipeline, I was not
successful is getting the SDL TSA task access to the file as the file is
created in a separate job and stage.
- As a result, the pipeline will keep the "SDT" job. Whenever there's a
solution with 1ES PT (for example, accept inputs), then we could likely
remove the SDT entirely as all tasks are handled by 1ES PT SDL analysis.
  - This PR does update the SDT tasks to the latest version.
- Example build where TSA is successful (otherwise the task runs only
merge happens on main branch):
https://dev.azure.com/accessibility-insights-private/Accessibility%20Insights%20(private)/_build/results?buildId=43592&view=results
- Creates a suppression file to skip CodeQL errors related to
`yarn-3.5.0.cjs`. Reference:
https://eng.ms/docs/cloud-ai-platform/devdiv/one-engineering-system-1es/1es-docs/1es-pipeline-templates/features/sdlanalysis/baselines#azure-devops-based-builds
- Adds an inline comment to suppress (what appears to be) a
false-positive warning on a markdown/escape function.
- I found that running windows and linux yielded the same build time, so
I defaulted to Windows given that the steps are platform agnostic and
1ES PT on Windows has more analysis tools at this time.
- Build performance:
   - 1ES PT runs in approximately 10-11 minutes
   - Previous pipeline ran in approximately 7-9 minutes 

##### Motivation

https://dev.azure.com/mseng/1ES/_workitems/edit/2033962

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)