fix: removing yaml dependency (#1650)
#### Details
Removed the `yaml` dependency and updated to a later version of the
`cosmiconfig` package that uses `js-yaml` instead

<!-- Usually a sentence or two describing what the PR changes -->

##### Motivation
The older version of `yaml` had a security vulnerability. Since the
dependency was introduced by `cosmiconfig`, opted to update this
package, which switched to the `js-yaml` parser

<!-- This can be as simple as "addresses issue #123" -->

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [N/A] Addresses an existing issue: Fixes #0000
- [N/A] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)