feat: Upgrade to axe-core 4.7.2 (#1769)
#### Details

This PR updates to axe-core 4.7.2 and ai-scan 2.3.1. The tests have all
been updated, including increasing the timeout of an e2e test that is
slightly slower due to additional redirect overheads that exist in this
scenario.

##### Motivation

Feature work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)