fix: Emit a warning when on an unsupported Windows platform (#1759)
Some Windows Server versions (Nano Server and Server Core) lack the
needed dependencies to run Chrome. This commit adds a check for this case
(based on the information in
https://github.com/microsoft/accessibility-insights-action/issues/1575#issuecomment-1528353206),
displaying a warning in the console when it occurs. Specifically, it
checks for the availability of the `Server-Media-Foundation` feature
(only available on server editions, not client) with non-installed
state.

This commit adds a new dependency
([wmi-client](https://www.npmjs.com/package/wmi-client)) to
`package.json`. In order to use this package and pass linting, we needed
to create a declaration file for its types. This is
`packages/ado-extension/src/wmic.d.ts`, which declares the
`"wmi-client"` `module` and its `wmic` `class`.

Addresses #1575.

Co-authored-by: madalynrose <madparker@microsoft.com>