Chore: Remove inline snapshots (#1825)
#### Details

Due to https://github.com/jestjs/jest/issues/14305, our inline snapshots
are causing problems when using `yarn test -u`. We discussed this
internally and decided that the simplest approach was to convert the
inline snapshots to traditional snapshots. For most cases, this was just
removing "Inline" from the method name and removing the string that
provided the snapshot. A small number of test cases were checking
multiple inline snapshots in the same test, so the approach there was to
build an array of the things that were previously checked individually,
then having a single snapshot for the test case.

I ran `yarn test -u` and `yarn format:fix` after making all of the
changes, just to make sure that we're ready for future changes.

It will probably be easiest to review this one commit at a time--each of
the 32 commits represents the conversion of a single test case.

##### Motivation

<!-- This can be as simple as "addresses issue #123" -->

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue:
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)